Hooked up add new route event - updated model to get everything matching, Changed Mode references back to type... it got too messy!, Changed route duration to double

+++ model/map/Route.java
@@ -7,7 +7,7 @@ public class Route {
 	private final String origin;
 	private final String destination;
 	private final String company;
-	private final int duration;//hours
+	private final double duration;//hours
 	private final Type type;
 	private double custPriceWeight;
 	private double custPriceVolume;
@@ -16,7 +16,7 @@ public class Route {
 	private boolean inService;
 
 
-	public Route(int id, String origin, String destination, String company, int duration, Type type, boolean inService,
+	public Route(int id, String origin, String destination, String company, double duration, Type type, boolean inService,
 			double custPriceWeight, double custPriceVolume, double transPriceWeight, double transPriceVolume) {
 		this.ID = id;
 		this.origin = origin;
@@ -49,7 +49,7 @@ public class Route {
 		return company;
 	}
 
-	public int getDuration() {
+	public double getDuration() {
 		return duration;
 	}
